split complete
