+++ include/parser.h
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/19 09:43:53 by spyun         #+#    #+#                 */
-/*   Updated: 2025/05/20 15:45:49 by spyun         ########   odam.nl         */
+/*   Updated: 2025/05/20 15:49:45 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
