+++ src/common/free_memory.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   free_utils.c                                       :+:    :+:            */
+/*   free_memory.c                                      :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/07 11:34:24 by spyun         #+#    #+#                 */
-/*   Updated: 2025/05/19 16:03:30 by spyun         ########   odam.nl         */
+/*   Updated: 2025/05/20 15:26:06 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
