+++ src/parser/file/file_parser.c
@@ -1,0 +1,44 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        ::::::::            */
+/*   file_parser.c                                      :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: spyun <spyun@student.codam.nl>               +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2025/05/08 08:55:27 by spyun         #+#    #+#                 */
+/*   Updated: 2025/05/20 15:05:48 by spyun         ########   odam.nl         */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "parser.h"
+
+int	parse_file(char *filename, t_game *game)
+{
+	int	fd;
+	int	ret;
+
+	fd = open(filename, O_RDONLY);
+	if (fd == -1)
+		return (perror("Error opening file"), -1);
+	ret = 0;
+	return (parse_map_content(game, fd, ret));
+}
+
+int	parse_map(int fd, t_game *game, char *first_line)
+{
+	t_list	*map_lines;
+
+	if (!has_valid_first_line(first_line))
+		return (free(first_line), -1);
+	if (!check_all_elements_set(&game->asset))
+		return (free(first_line), -1);
+	map_lines = NULL;
+	if (read_map_lines(fd, &map_lines, first_line) == -1)
+		return (-1);
+	if (build_map_array(&game->map, map_lines) == -1)
+		return (ft_lstclear(&map_lines, free), -1);
+	ft_lstclear(&map_lines, free);
+	if (!validate_map(&game->map, &game->player))
+		return (free_map(&game->map), -1);
+	return (1);
+}
