+++ src/parser/map/map_borders.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   map_borders_validator.c                            :+:    :+:            */
+/*   map_borders.c                                      :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/13 11:42:12 by spyun         #+#    #+#                 */
-/*   Updated: 2025/05/19 09:48:10 by spyun         ########   odam.nl         */
+/*   Updated: 2025/05/20 15:08:42 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
