+++ src/parser/map/map_spaces.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   map_spaces_validator.c                             :+:    :+:            */
+/*   map_spaces.c                                       :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/13 11:43:41 by spyun         #+#    #+#                 */
-/*   Updated: 2025/05/19 16:33:38 by spyun         ########   odam.nl         */
+/*   Updated: 2025/05/20 15:08:45 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
