--- src/parser/parse_elements.c
@@ -1,201 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        ::::::::            */
-/*   parse_elements.c                                   :+:    :+:            */
-/*                                                     +:+                    */
-/*   By: spyun <spyun@student.codam.nl>               +#+                     */
-/*                                                   +#+                      */
-/*   Created: 2025/05/07 10:16:05 by spyun         #+#    #+#                 */
-/*   Updated: 2025/05/20 14:40:43 by spyun         ########   odam.nl         */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "parser.h"
-
-static void	ft_handle_char(char *trimmed, int *i, int *j, int *in_space)
-{
-	if (ft_isspace(trimmed[*i]))
-	{
-		if (!(*in_space))
-		{
-			trimmed[(*j)++] = ' ';
-			*in_space = 1;
-		}
-	}
-	else
-	{
-		trimmed[(*j)++] = trimmed[*i];
-		*in_space = 0;
-	}
-	(*i)++;
-}
-
-static char	*ft_trim_and_compact(char *str)
-{
-	char	*trimmed;
-	int		i;
-	int		j;
-	int		in_space;
-
-	trimmed = ft_strtrim(str, " \t\n\v\f\r");
-	if (!trimmed)
-		return (NULL);
-	i = 0;
-	j = 0;
-	in_space = 0;
-	while (trimmed[i])
-		ft_handle_char(trimmed, &i, &j, &in_space);
-	trimmed[j] = '\0';
-	return (trimmed);
-}
-
-static int	validate_texture(char **path, char **split)
-{
-	int	ret;
-
-	ret = 0;
-	*path = ft_strdup(split[1]);
-	if (!(*path))
-		ret = -1;
-	else if (!check_texture_path(*path))
-	{
-		free(*path);
-		*path = NULL;
-		ret = -1;
-	}
-	ft_free_strarr(split);
-	return (ret);
-}
-
-static int	parse_texture_path(char *line, char **path)
-{
-	char	*trimmed;
-	char	**split;
-
-	if (*path != NULL)
-		return (ft_putendl_fd("Error: Duplicate texture path", 2), -1);
-	trimmed = ft_trim_and_compact(line);
-	if (!trimmed)
-		return (-1);
-	split = ft_split(trimmed, ' ');
-	free(trimmed);
-	if (!split)
-		return (-1);
-	if (!split[0] || !split[1] || split[2])
-		return (ft_free_strarr(split), -1);
-	return (validate_texture(path, split));
-}
-
-static int	parse_single_value(char **str, int *value)
-{
-	int		j;
-	bool	negative;
-
-	while (**str && (**str == ' ' || **str == ','))
-		(*str)++;
-	negative = false;
-	if (**str == '-')
-	{
-		negative = true;
-		(*str)++;
-	}
-	else if (**str == '+')
-		(*str)++;
-	j = 0;
-	*value = 0;
-	while (**str && ft_isdigit(**str))
-	{
-		*value = *value * 10 + (**str - '0');
-		(*str)++;
-		j++;
-	}
-	if (negative)
-		*value = -*value;
-	return (j);
-}
-
-static int	parse_rgb_values(char *str, int values[3])
-{
-	int	i;
-	int	j;
-
-	i = 0;
-	while (i < 3)
-	{
-		j = parse_single_value(&str, &values[i]);
-		if (j == 0)
-			return (-1);
-		i++;
-	}
-	return (0);
-}
-
-static int	set_color_values(t_color *color, int values[3])
-{
-	color->r = values[0];
-	color->g = values[1];
-	color->b = values[2];
-	if (!check_color_value(color->r, color->g, color->b))
-	{
-		color->r = -1;
-		color->g = -1;
-		color->b = -1;
-		return (-1);
-	}
-	return (0);
-}
-
-static int	extract_rgb_values(char *line, t_color *color)
-{
-	int		values[3];
-	char	*str;
-
-	if (color->r >= 0)
-		return (ft_putendl_fd("Error: Duplicate color value", 2), -1);
-	str = line;
-	while (*str && *str != ' ')
-		str++;
-	while (*str && *str == ' ')
-		str++;
-	if (parse_rgb_values(str, values) == -1)
-		return (-1);
-	return (set_color_values(color, values));
-}
-
-static int	parse_color_value(char *line, t_color *color)
-{
-	char	*trimmed;
-	int		ret;
-
-	trimmed = ft_trim_and_compact(line);
-	if (!trimmed)
-		return (-1);
-	ret = extract_rgb_values(trimmed, color);
-	free(trimmed);
-	return (ret);
-}
-
-int	parse_element(char *line, t_asset *asset)
-{
-	char	*trimmed;
-	int		ret;
-
-	trimmed = ft_trim_and_compact(line);
-	if (!trimmed)
-		return (-1);
-	ret = -1;
-	if (ft_strncmp(trimmed, "NO ", 3) == 0)
-		ret = parse_texture_path(line, &asset->no_path);
-	else if (ft_strncmp(trimmed, "SO ", 3) == 0)
-		ret = parse_texture_path(line, &asset->so_path);
-	else if (ft_strncmp(trimmed, "WE ", 3) == 0)
-		ret = parse_texture_path(line, &asset->we_path);
-	else if (ft_strncmp(trimmed, "EA ", 3) == 0)
-		ret = parse_texture_path(line, &asset->ea_path);
-	else if (ft_strncmp(trimmed, "F ", 2) == 0)
-		ret = parse_color_value(line, &asset->floor);
-	else if (ft_strncmp(trimmed, "C ", 2) == 0)
-		ret = parse_color_value(line, &asset->ceiling);
-	free(trimmed);
-	return (ret);
-}
